Changed the response format of the sign up endpoint.
Old format was the token as a string, new format is a JSON object with fields for the names, email, token and the user's subscription.
The user's subscription is represented with an object, which is null, for newly created accounts.

Token includes a subject claim with the db id of the user, for future authorization concerns.